Introduced array.dedup.* arguments to deduplicate all arrays.
This extends the deduplication machinery from DelayedArray seeds to all arrays
that pass through alabaster.matrix's saveObject methods. The aim is to allow
deduplication of arrays outside of DelayedArrays, e.g., if a SE contains a
non-delayed array and a DelayedArray wrapping the first array, we can now
deduplicate the seed in the latter when a dedup session is supplied for the SE.

This change involves migrating many of the deduplication functions over to
alabaster.base for more general use. We also add a DelayedArray.force.external=
argument to make it easier to deduplicate DelayedArray seeds.